Literature client functions are now in literature.__init__
